+++ backend/api/tests.py
@@ -1,5 +1,5 @@
 from django.test import TestCase
-from backend.api.views import Hex2ViewSet
+from backend.api.api_views.hex2_view_set import Hex2ViewSet
 
 
 class TestAPI(TestCase):
