+++ backend/urls.py
@@ -1,6 +1,9 @@
 from django.urls import path
 
-from .api.views import index_view, Hex2ViewSet, InfoViewSet, CreateViewSet
+from .api.views import index_view
+from .api.api_views.create_view_set import CreateViewSet
+from .api.api_views.info_view_set import InfoViewSet
+from .api.api_views.hex2_view_set import Hex2ViewSet
 
 handler404 = "backend.api.views.handler404_redirect"
